requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
numpy==2.2.6
openai==1.99.9
opencv-python==4.12.0.88
pydantic_core==2.33.2
pydantic==2.11.7
python-docx==1.2.0
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.1
tqdm==4.67.1
typing_extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
boto3==1.35.50
requests==2.32.3